requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Pillow==10.4.0
piexif==1.1.3
# XMP支持库
python-xmp-toolkit==2.0.1
# 备用元数据库，提供更强健的fallback支持
ExifRead==3.0.0
